renamed utility folder to private_values
to make more understandable its purpose
